Support wget 1.17's usage of If-Modified-Since with -N/--timestamping
Consider a dphys-config.list file like this:

foo.conf:/etc/foo-23/
foo.conf:/etc/foo-42/

dphys-config tries twice to download the file "foo.conf" from the
configuration files repository with wget. In the past, "wget -N"
noticed that the file is already there ("Server file no newer than
local file `foo.conf' -- not retrieving"), but still logged a "200 OK"
response code ("HTTP request sent, awaiting response... 200 OK"). So
dphys-config saw the "200" and was happy with it.

But since version 1.17, wget uses If-Modified-Since headers with the
"-N" respectively "--timestamping" option and reports a "304 Not
Modified" instead of the previous "Server file no newer than local
file `foo.conf' -- not retrieving" message. The HTTP code "304"
confuses dphys-config as of now and it considers it an error and
aborts.

Closes: #819557